Move periodical util to `pkg/util`
